Turn into a distributable Python project
* Move juniper-vnp.py to junipervpn/vpn.py and tncc.py to
  junipervpn/tncc.py, thereby making them modules in the junipervpn
  package.

* Provide a setup.py script, suitable for use as "pip install ." with
  the necessary list of dependencies, and the CLI tool entry point.
  Installing this way will provide a "juniper-vpn.py" binary in a
  standard location in the PATH.

* Add compatibility shim for backward compatibility with the old
  installation method. It includes a warning as using that shim is not
  recommended at all.
